fix: implement homepage
This implements the v0.5 of the user facing homepage in Yari.
To take this over the finish line I need assistance from @peterbe with regards to the hacks blog feed.

Part of #2438
