+++ client/src/author-dashboard/viewed-documents.scss
