+++ client/src/author-dashboard/viewed-documents.tsx
