+++ client/src/homepage/index.scss
@@ -1,0 +1,32 @@
+@import "~@mdn/minimalist/sass/mixins/utils";
+@import "~@mdn/minimalist/sass/vars/color-palette";
+@import "~@mdn/minimalist/sass/vars/layout";
+
+.hero {
+  background-color: $primary-300;
+  padding: $base-spacing;
+}
+
+.hero-content {
+  max-width: $max-width-default;
+
+  h1 {
+    background-color: $neutral-100;
+    color: $text-color-inverted;
+    font-weight: normal;
+    margin-bottom: $base-spacing / 2;
+    text-align: center;
+  }
+
+  ul {
+    @include reset-list();
+
+    a {
+      border-bottom: 2px solid $neutral-100;
+      color: $neutral-100;
+      display: inline-block;
+      padding: ($base-spacing / 2) 0;
+      width: 100%;
+    }
+  }
+}
